Don't write header multiple times

+++ src/stream_tachograph.py
@@ -2,6 +2,7 @@
 
 import random
 import sys
+import os
 import pandas as pd
 from time import sleep
 
@@ -18,7 +19,8 @@ OUT_PATH = "tmp/{}.csv".format(DRIVER_ID)
 df = pd.read_csv(DATA_PATH)
 
 # Write header
-df.loc[[0]].to_csv(OUT_PATH, mode='a', header=True, index=False)
+if not os.path.isfile(OUT_PATH):
+    df.loc[[0]].to_csv(OUT_PATH, mode='a', header=True, index=False)
 
 for i in range(1, len(df)):
     df.loc[[i]].to_csv(OUT_PATH, mode='a', header=False, index=False)
